Extracted Segmentation Pair Class
